+++ packages/react-native/src/components/Accordion/Accordion.tsx
@@ -110,7 +110,7 @@ const Accordion = forwardRef<View, AccordionProps>(
 let accordionItemCounter = 0;
 
 export const AccordionItem = forwardRef<View, AccordionItemProps>(
-  ({ children, value, title, expanded, triggerContent, ...props }, ref) => {
+  ({ children, value, title, expanded, triggerContent, testID, ...props }, ref) => {
     if (!children) {
       return null;
     }
@@ -143,7 +143,7 @@ export const AccordionItem = forwardRef<View, AccordionItemProps>(
         ) : (
           <>
             <AccordionHeader>
-              <AccordionTrigger isExpanded={expanded}>
+              <AccordionTrigger isExpanded={expanded} testID={testID}>
                 {({ isExpanded }: { isExpanded: boolean }) => {
                   return (
                     <>