+++ packages/react/src/components/Checkbox/Checkbox.tsx
@@ -1,20 +1,18 @@
 import * as React from 'react';
-
 import clsx from 'clsx';
-
 import type { CheckboxProps } from './Checkbox.props';
 import { Flex } from '../Flex/Flex';
 import { Label } from '../Label/Label';
-import { useFormFieldGroup } from '../FormFieldGroup/FormFieldGroup.context';
 import { HelperText } from '../HelperText/HelperText';
 import type { ElementRef } from 'react';
 import { TickSmallIcon } from '@utilitywarehouse/hearth-react-icons';
 import { withGlobalPrefix } from '../../helpers/with-global-prefix';
-import { useCheckboxGroupBase } from '../CheckboxGroupBase/CheckboxGroupBase.context';
 import { useIds } from '../../hooks/use-ids';
 import { marginPropDefs } from '../../props/margin.props';
 import { extractProps } from '../../helpers/extract-props';
 import { Checkbox as RadixCheckbox } from 'radix-ui';
+import { useCheckboxGroup } from '../CheckboxGroup/CheckboxGroup.context';
+import { useFormGroupBase } from '../FormGroupBase/FormGroupBase.context';
 
 const componentName = 'Checkbox';
 const componentClassName = withGlobalPrefix(componentName);
@@ -34,8 +32,8 @@ export const Checkbox = React.forwardRef<CheckboxElement, CheckboxProps>((props,
     ...checkboxProps
   } = extractProps(props, marginPropDefs);
   const { id, labelId, helperTextId } = useIds({ providedId, prefix: 'checkbox' });
-  const context = useFormFieldGroup();
-  const checkboxContext = useCheckboxGroupBase();
+  const context = useFormGroupBase();
+  const checkboxContext = useCheckboxGroup();
   const checked = checkboxContext?.value?.includes(value);
   const hasGroupHelperText = context?.hasGroupHelperText;
   const ariaDescribedby = context ? context['aria-describedby'] : '';