+++ packages/react/src/components/CheckboxTile/CheckboxTile.tsx
@@ -5,14 +5,14 @@ import clsx from 'clsx';
 import { CheckboxTileProps } from './CheckboxTile.props';
 import { withGlobalPrefix } from '../../helpers/with-global-prefix';
 import type { ElementRef } from 'react';
-import * as RadixCheckbox from '@radix-ui/react-checkbox';
 import { useIds } from '../../hooks/use-ids';
-import { useFormFieldGroup } from '../FormFieldGroup/FormFieldGroup.context';
 import { TickSmallIcon } from '@utilitywarehouse/hearth-react-icons';
-import { useCheckboxGroupBase } from '../CheckboxGroupBase/CheckboxGroupBase.context';
 import { Label } from '../Label/Label';
 import { HelperText } from '../HelperText/HelperText';
 import { Flex } from '../Flex/Flex';
+import { Checkbox as RadixCheckbox } from 'radix-ui';
+import { useCheckboxGroup } from '../CheckboxGroup/CheckboxGroup.context';
+import { useFormGroupBase } from '../FormGroupBase/FormGroupBase.context';
 
 const componentName = 'CheckboxTile';
 const componentClassName = withGlobalPrefix(componentName);
@@ -35,8 +35,8 @@ export const CheckboxTile = React.forwardRef<CheckboxTileElement, CheckboxTilePr
     ref
   ) => {
     const { id, labelId, helperTextId } = useIds({ providedId, prefix: 'radio' });
-    const context = useFormFieldGroup();
-    const checkboxContext = useCheckboxGroupBase();
+    const context = useFormGroupBase();
+    const checkboxContext = useCheckboxGroup();
     const checked = checkboxContext?.value?.includes(value);
     const hasGroupHelperText = context?.hasGroupHelperText;
     const ariaDescribedby = context ? context['aria-describedby'] : '';