+++ packages/react/src/components/FormGroupBase/Legend.props.ts
@@ -1,4 +1,4 @@
 import { MarginProps } from '../../props/margin.props';
 import { ComponentPropsWithout, RemovedProps } from '../../types/component-props';
 
-export type FieldsetProps = ComponentPropsWithout<'fieldset', RemovedProps> & MarginProps;
+export type LegendProps = ComponentPropsWithout<'legend', RemovedProps> & MarginProps;